backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
mysqlclient==2.2.0
python-dotenv==1.0.0
cloudinary==1.38.0
Pillow==10.2.0
